feat: refactor rooms code
