+++ packages/api/src/common/constants/database.ts
@@ -1,0 +1,1 @@
+export const DATABASE_CONNECTION_KEY = 'DATABASE_CONNECTION';
