+++ packages/api/src/common/constants/models/room.ts
@@ -1,0 +1,2 @@
+export const DB_ROOM_MODEL_KEY = 'ROOM_MODEL';
+export const DB_ROOM_MODEL_NAME = 'Room';
