+++ packages/api/src/common/types/room.ts
@@ -1,4 +1,4 @@
-import { Document } from "mongoose";
+import { Document } from 'mongoose';
 
 export interface Room extends Document {
   readonly name: string;
