+++ packages/api/tsconfig.json
@@ -19,6 +19,7 @@
     "noFallthroughCasesInSwitch": false,
     "paths": {
       "@/auth/*": ["./src/auth/*"],
+      "@/common/*": ["./src/common/*"],
       "@/database/*": ["./src/database/*"],
       "@/rooms/*": ["./src/rooms/*"],
       "@/types/*": ["../types/*"]
