--- packages/types/constants/database.ts
@@ -1,1 +1,0 @@
-export const DATABASE_CONNECTION_KEY = "DATABASE_CONNECTION";
