--- packages/types/constants/models/room.ts
@@ -1,2 +1,0 @@
-export const DB_ROOM_MODEL_KEY = "ROOM_MODEL";
-export const DB_ROOM_MODEL_NAME = "Room";
